Add Gitter Badge (#2)
This commit is performed as a branch and PR (pull request) to the default branch instead of simply merging it. This illustrates the work flow intended for work on LWFBrook90.jl (git branch model).

Apperently, this procedure (PR instead of merge) allows to more easily keep track of your work in issues insted of simple branches.

E.g. in the following way. Part of this commit fixes issue/PR #1.
The commit additionally includes a small modification in the code making looping clearer.