Include Hammel-2001 simulations as integration tests and save only 1x daily
Including the infiltration simulation form Hammel et al. (2001) as
integration tests. The tests compare the LWFBrook90.jl implementation
with the solutions generated with LWFBrook90R 0.4.3 that are stored as
*.csv's. Note that given the nature of the simulation, many of the
aboveground processes are deactivated and it's a comparatively simple
infiltration test. These tests do thus not cover the aboveground
compartments.

Further note that this commit changes the saving behvior of
LWFBrook90.jl. Before this change the solution was stored twice, once
before applying the daily callback and once after. Resulting in twice
the number of days as output. The change was needed for easy comparison
with the hardcoded solutions from LWFBrook90R.

Note: consider undoing this change and keep saving twice. Then in the
comparison of the tests be more careful which values we use to compare.