feat: Decrease initial time step similar to R pkg
Time step for preliminary calculation of water fluxes (VRFLI) is reduced
to be at most the remaining time until the end of the day. This brings
the code closer to the implementation in LWFBrook90R.

This change might modify simulation results slightly. However, this
changes only touches the preliminiray calculation, which is anyway
refined in a secondary step.