Include Hammel-2001 simulations as integration tests (#27)
Hammel-2001 simulations are included as integration tests. One comparison is made versus the solution generated with `LWFBrook90R` using the same discretization as with `LWFBrook90.jl`. The other comparison is made versus a single Hydrus simulation (using a homogenous discretization of 20mm).

The branch was merged using the squash and merge strategy on github.com. Below are the commit titles and descriptions of what is included.

* chore: Extend README-for-Developers.md with git commit and julia docs/test tips

* Include Hammel-2001 simulations as integration tests and save only 1x daily

Including the infiltration simulation form Hammel et al. (2001) as
integration tests. The tests compare the LWFBrook90.jl implementation
with the solutions generated with LWFBrook90R 0.4.3 that are stored as
*.csv's. Note that given the nature of the simulation, many of the
aboveground processes are deactivated and it's a comparatively simple
infiltration test. These tests do thus not cover the aboveground
compartments.

Further note that this commit changes the saving behvior of
LWFBrook90.jl. Before this change the solution was stored twice, once
before applying the daily callback and once after. Resulting in twice
the number of days as output. The change was needed for easy comparison
with the hardcoded solutions from LWFBrook90R.

Note: consider undoing this change and keep saving twice. Then in the
comparison of the tests be more careful which values we use to compare.

* test: Split tests into unit-,integration-,and regression

* chore: Add example scripts in folder "examples/"

* Include raw files to reproduce Hammel-2001 with Hydrus and LWFBrook90R

* test: Clean up unnecessary copies in integration tests

* test: Fix test failure due to benchmark test

* test: Add Hammel-2001 test vs Hydrus1D

Adding comparison of θ of Hammel-2001 simulations relative to Hydrus1D
solutions in the test set.

* refactor: Make callback function accessible for perf. optim

This commit makes callback function accessible outside of the `solve()`
in order to analyze its performance more easily.