perf: Improve allocated memory in KPT
This commit improves the allocated memory in some functions from the
module KPT. The goal is to improve performance with these modifications.

A next step, however, would be to simply profile the code and see where
it spends time instead of trying to optimize everything.

As immediate next step one could try to improve type stability of f and
cb.