test: Include BEA-integration test vs R pkg
This commit includes simulations based on BEA-2016, that have been run
with both, LWFBrook90.jl and LWFBrook90R. Solutions in terms of below-
ground (θ and ψ for certain layers) and aboveground (states: INTR, INTS,
SNOW, and GWAT; [not CC or SNOWLQ as they are not output by LWFBrook90R]).

LWFBrook90R was run with different discretiations NLAYER (7,14,21,70).
LWFBrook90.jl just with NLAYER == 7.