feat: Simplify API by defining solve_LWFB90() and fix adaptivity
Defines solve_LWFB90(u0, tspan, p) and gets rid of define_LWFB90_ODE().

This allows to define internally the RHS, the callbacks, the
unstable_check_function, the internalnorm as well as all the solver parameter
that need to be defined to solve a system. By internalizing it we can change it
later without the user needing to know.

Public (user-facing) API are now only the u0, tspan and p. (Note that p might
benefit from simplifications).