test: Fix integration and regression tests Also when running on MacBook integration and regression tests do not give exactly the same results depending on whether they are run in the interactive REPL or from Pkg() with `test`.
This is strange. It should not be any Random seed as the testsuite
should automatically fix this. Nevertheless, based on the source:
https://www.juliabloggers.com/testing-julia/ from April 2022 we still
give it a try with Random.seed!().

An alternative option, that could also explain the failing on the CI,
could be that the seed only works for the same versions of Julia, and
otherwise, one would need to go about Random.seed!() (same source as above).